+++ starts-maven-plugin/src/main/java/edu/illinois/starts/jdeps/ImpactedMojo.java
@@ -8,8 +8,7 @@ import edu.illinois.starts.helpers.RTSUtil;
 import edu.illinois.starts.helpers.Writer;
 import edu.illinois.starts.helpers.ZLCHelper;
 import edu.illinois.starts.plugin.StartsPluginException;
-import edu.illinois.starts.plugin.StartsPluginImpactedGoal;
-import edu.illinois.starts.plugin.StartsPluginMavenGoal;
+import edu.illinois.starts.plugin.buildsystem.StartsPluginMavenGoal;
 import edu.illinois.starts.util.Logger;
 import edu.illinois.starts.util.Pair;
 import edu.illinois.starts.util.Result;
@@ -32,7 +31,7 @@ import java.util.logging.Level;
  */
 @Mojo(name = "impacted", requiresDirectInvocation = true, requiresDependencyResolution = ResolutionScope.TEST)
 @Execute(phase = LifecyclePhase.TEST_COMPILE)
-public class ImpactedMojo extends DiffMojo implements StartsPluginMavenGoal, StartsPluginImpactedGoal {
+public class ImpactedMojo extends DiffMojo implements StartsPluginMavenGoal {
     /**
      * Set this to "true" to update test dependencies on disk. The default value of "false"
      * is useful for "dry runs" where one may want to see the diff without updating
