+++ starts-plugin-util/pom.xml
@@ -46,11 +46,6 @@
         </dependency>
         <dependency>
             <groupId>org.gradle</groupId>
-            <artifactId>gradle-tooling-api</artifactId>
-            <version>3.4.1</version>
-        </dependency>
-        <dependency>
-            <groupId>org.gradle</groupId>
             <artifactId>gradle-base-services</artifactId>
             <version>3.4.1</version>
         </dependency>
