--- starts-plugin-util/src/main/java/edu/illinois/starts/plugin/StartsPluginDiffGoal.java
@@ -1,4 +1,0 @@
-package edu.illinois.starts.plugin;
-
-public interface StartsPluginDiffGoal extends StartsPluginBaseGoal {
-}
