+++ starts-plugin-util/src/main/java/edu/illinois/starts/plugin/goals/StartsPluginImpactedGoal.java
@@ -1,4 +1,4 @@
-package edu.illinois.starts.plugin;
+package edu.illinois.starts.plugin.goals;
 
 public interface StartsPluginImpactedGoal extends StartsPluginBaseGoal {
